Extract tracing protocols and telemetry test support
Move the PEP 544 Span and Tracer protocols out of the tracing adapter into
cuprum/adapters/_tracing_protocols.py, and lift the shared telemetry
behaviour-test scaffolding into tests/behaviour/_telemetry_adapter_support.py
so each module stays under the size ceiling.
